新增debug trace功能example
